src/chat/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
transformers==4.35.0
torch==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiohttp==3.9.0
cryptography==41.0.5
pyOpenSSL==23.3.0
certifi==2023.11.17
email-validator==2.1.0.post1
python-magic==0.4.27
safety==2.3.5
bandit==1.7.5
